# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API - Max CPFs",
        "description": "Documentação do Projeto API Max CPFs",
        "contact": {
            "name": "João Paulo Pinheiro",
            "email": "joaopinheiro.ti@gmail.com.br"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://maxcpf.local/api/v1",
            "description": "Ambiente de Desenvolvimento"
        },
        {
            "url": "http://maxcpf.prod/api/v1",
            "description": "Ambiente de Produção"
        }
    ],
    "paths": {
...
        "/cpf/{cpf}": {
            "get": {
                "tags": [
                    "CPFs"
                ],
                "summary": "Checar CPF.",
                "operationId": "App\\Http\\Controllers\\CpfController::show",
                "parameters": [
                    {
                        "name": "cpf",
                        "in": "path",
                        "description": "Informe o CPF",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "showUser",
                        "in": "query",
                        "description": "Exibir Usuário",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": "false"
                    },
                    {
                        "name": "showRole",
                        "in": "query",
                        "description": "Exibir os Perfis do Usuário",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": "false"
                    },
                    {
                        "name": "showPermission",
                        "in": "query",
                        "description": "Exibir as Permissões do Perfil do Usuário",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": "false"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resposta Operacional Normal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/CpfContent"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Sem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CPFs"
                ],
                "summary": "Remover CPF.",
                "operationId": "App\\Http\\Controllers\\CpfController::destroy",
                "parameters": [
                    {
                        "name": "cpf",
                        "in": "path",
                        "description": "Informe o CPF",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resposta Operacional Normal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiDelete"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Sem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/permissions/{id}": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Listar permissão por ID.",
                "operationId": "App\\Http\\Controllers\\PermissionController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da Permissão",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resposta Operacional Normal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/PermissionContent"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Sem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Atualizar permissão.",
                "operationId": "App\\Http\\Controllers\\PermissionController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da Permissão",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "display_name",
                                    "description"
                                ],
                                "properties": {
                                    "display_name": {
                                        "description": "Nome de Exibição",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Descrição",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resposta Operacional Normal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/PermissionContent"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Sem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Deletar permissão.",
                "operationId": "App\\Http\\Controllers\\PermissionController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da Permissão",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resposta Operacional Normal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiDelete"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Sem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/roles/{id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Listar perfil por ID.",
                "operationId": "App\\Http\\Controllers\\RoleController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Perfil",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resposta Operacional Normal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/RoleContent"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Sem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Atualizar perfil.",
                "operationId": "App\\Http\\Controllers\\RoleController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Perfil",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "permissions[]",
                        "in": "query",
                        "description": "Permissões do Perfil",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "display_name",
                                    "description"
                                ],
                                "properties": {
                                    "display_name": {
                                        "description": "Nome de Exibição",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Descrição",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resposta Operacional Normal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/RoleContent"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Sem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Deletar perfil e permissões em cascata.",
                "operationId": "App\\Http\\Controllers\\RoleController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Perfil",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resposta Operacional Normal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiDelete"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Sem permissão de acesso"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/auth/register": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Registrar uma nova conta no sistema.",
                "operationId": "App\\Http\\Controllers\\UserController::register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "E-mail",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Senha",
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "Confirmar Senha",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resposta Operacional Normal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/UserContent"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Sem permissão de acesso"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "title": "Retorno padrão",
                "description": "Entidade de resposta, resultado da resposta usa essa estrutura uniformemente",
                "properties": {
                    "success": {
                        "description": "Status da resposta",
                        "type": "boolean"
                    },
                    "code": {
                        "description": "Código da resposta",
                        "type": "integer"
                    },
                    "message": {
                        "description": "Mensagem da resposta",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ApiDelete": {
                "title": "Retorno padrão",
                "description": "Entidade de resposta, resultado da resposta usa essa estrutura uniformemente",
                "properties": {
                    "success": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "CpfContent": {
                "title": "CPF",
                "description": "CPF",
                "properties": {
                    "Content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CpfResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "CpfResource": {
                "title": "CPF",
                "description": "CPF",
                "properties": {
                    "cpf": {
                        "description": "CPF",
                        "type": "string"
                    },
                    "createdAt": {
                        "description": "createdAt",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "PermissionContent": {
                "title": "Auth Permissions",
                "description": "Permissões de Usuário",
                "properties": {
                    "Content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PermissionResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "RoleContent": {
                "title": "Auth Roles",
                "description": "Perfil de Usuário",
                "properties": {
                    "Content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RoleResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "UserContent": {
                "title": "User",
                "description": "Usuário",
                "properties": {
                    "Content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserResource"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Informe o token JWT para endpoits seguros",
                "name": "JWT bearer",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}